+++ packages/core/src/RPCMessages/BladeReauthenticate.ts
@@ -1,4 +1,4 @@
-import { makeRPCRequest } from './index'
+import { makeRPCRequest } from './helpers'
 import { BladeMethod } from '../utils/constants'
 
 type BladeReauthenticateParams = { project: string; jwt_token: string }
