+++ packages/core/src/RPCMessages/helpers.ts
@@ -1,0 +1,30 @@
+import { v4 as uuidv4 } from 'uuid'
+
+type MakeRPCRequestParams = {
+  method: string // TODO: use enum
+  params: {
+    // TODO: use list of types?
+    [key: string]: any
+  }
+}
+export const makeRPCRequest = (params: MakeRPCRequestParams) => {
+  return {
+    jsonrpc: '2.0' as const,
+    id: uuidv4(),
+    ...params,
+  }
+}
+
+type MakeRPCResponseParams = {
+  id: string
+  result: {
+    // TODO: use list of types?
+    [key: string]: any
+  }
+}
+export const makeRPCResponse = (params: MakeRPCResponseParams) => {
+  return {
+    jsonrpc: '2.0' as const,
+    ...params,
+  }
+}
